//Mahesh

+++ src/main/java/com/monginis/ops/controller/OpeneningStockController.java
@@ -118,7 +118,7 @@ public class OpeneningStockController {
 			String stockQty = request.getParameter("stockQty" + detailList.get(i).getItemId());
 			System.out.println("new qty " + stockQty);
 
-			detailList.get(i).setRegOpeningStock(Integer.parseInt(stockQty));
+			detailList.get(i).setRegOpeningStock(Float.parseFloat(stockQty));
 
 		}
